Reorganized and removed dot imports
